requirements.txt
google-cloud-datastore==1.4.0
google-api-python-client